Replace Parse.optional() with a validation function
Optional is really a type term. The problem it was trying to solve was
how to parse an untrustworthy query string. This PR replaces optional
with a required validation function, forcing the dev to explicitly deal
with the case where the data comes in bad.

The input arg of the validation is either unknown, undefined or a
primitive. The user can then deal with the unknown/undefined case and
return the required type. This means they can also check against unions
and have stricter types that just primitives.